test: move all devnet code under internal/test/devnet
